+++ src/bp/ui-studio/src/web/views/Config/index.tsx
@@ -285,7 +285,7 @@ class ConfigView extends Component<Props, State> {
     }
 
     try {
-      const res = await axios.post(`bots/${this.props.bot.id}/media`, data, {
+      const res = await axios.post(`studio/${this.props.bot.id}/media`, data, {
         ...axiosConfig,
         headers: { 'Content-Type': 'multipart/form-data' }
       })
